Set ROOT to top_builddir
Building with MSYS2 was failing due to the command line for linking glib
DLL exceeding the sh.exe limit of 8192 bytes. When sh.exe from MSYS2 was
present in PATH, gmake would use it and run a truncated command.

Using relative filenames shortens that command line to about 5k
regardless of the source directory location.